# Faster Day4_Part2.py