Before: 120.84 seconds
% time python3 Day4_Part2.py
7185540
python3 Day4_Part2.py  120.84s user 0.29s system 99% cpu 2:01.46 total

After: 0.07 seconds
% time python3 Day4_Part2.py                   
7185540
python3 Day4_Part2.py  0.07s user 0.02s system 90% cpu 0.098 total